refactor(pocha): move statusTone.ts from _shared to utils
Consolidates dashboard status-tone helpers under the existing utils/
folder; deletes the now-empty _shared directory.
